+++ src/composable/useWorker.ts
@@ -141,6 +141,8 @@ function useHandlers(mp: MessagePort) {
 				events.emit("twitch_emote_set_data", out);
 				break;
 			}
+			case "EMOTE_SET_UPDATED":
+				events.emit("emote_set_updated", data as TypedWorkerMessage<"EMOTE_SET_UPDATED">);
 		}
 	});
 }
@@ -181,7 +183,8 @@ type WorkletEventName =
 	| "entitlement_created"
 	| "entitlement_deleted"
 	| "static_cosmetics_fetched"
-	| "twitch_emote_set_data";
+	| "twitch_emote_set_data"
+	| "emote_set_updated";
 
 type WorkletTypedEvent<EVN extends WorkletEventName> = {
 	ready: object;
@@ -190,6 +193,7 @@ type WorkletTypedEvent<EVN extends WorkletEventName> = {
 	entitlement_deleted: Pick<SevenTV.Entitlement, "id" | "kind" | "ref_id" | "user_id">;
 	static_cosmetics_fetched: Pick<TypedWorkerMessage<"STATIC_COSMETICS_FETCHED">, "badges" | "paints">;
 	twitch_emote_set_data: SevenTV.EmoteSet;
+	emote_set_updated: TypedWorkerMessage<"EMOTE_SET_UPDATED">;
 }[EVN];
 
 export class WorkletEvent<T extends WorkletEventName> extends CustomEvent<WorkletTypedEvent<T>> {