+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Added support for Replies & Threads
 - Added support for Mod Icons
+- Added a message announcement when channel 7TV emotes are changed
+- Added an option to change the chat padding style (full-width or native)
 - The emote menu button will now take the form of the 7TV Logo
 - Fixed an issue which caused this changelog to become impossible to close if multiple tabs were open